--- server/controllers/user.controller.js
@@ -1,6 +1,0 @@
-import { getUser } from "../service/user.service.js"
-
-export const getall = async (req,res) => {
-    const userinfo = await getUser()
-    res.json(userinfo)
-}
