+++ server/routers/post.router.js
@@ -1,6 +1,8 @@
 import express from "express";
-import { getall } from "../controllers/user.controller.js";
+import { getall } from "../controllers/post.controller.js";
+import {post} from "../controllers/post.controller.js"
 import cloudinary from "../service/cloudniary.js";
+
 
 const router = express.Router();
 
@@ -11,22 +13,7 @@ router.get("/", async (req, res) => {
   res.json(data)
 });
 
-router.post("/", async (req, res) => {
-  try {
-    const data = req.body;
-    const fileStr = req.body.imageUrl;
-    const uploadedResponse = await cloudinary.uploader.upload(fileStr, {
-      folder: "profile_pic",
-      width: "800"
-    });
-    data.imageUrl = uploadedResponse.url;
-    console.log(data);
-    res.json(data);
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: "An error occurred while uploading the image" });
-  }
-});
+router.post("/", post)
 
 router.put("/", async (req, res) => {
   try {
